Fix #41 and update deleted forms method

+++ studio/views/subject_views.py
@@ -9,6 +9,7 @@ from django.views.generic import CreateView, ListView, UpdateView
 from sweetify import sweetify
 
 from alumnica_model.models import Tag, Subject, Ambit
+from alumnica_model.models.content import Image
 from studio.forms.subject_forms import SubjectForm, BaseImageFormset, ImageForm, UpdateSubjectForm
 
 
@@ -177,7 +178,8 @@ class UpdateSubjectView(LoginRequiredMixin, UpdateView):
 
             for deleted_form in formset.deleted_forms:
                 object_to_delete = deleted_form.save(commit=False)
-                object_to_delete.delete()
+                if object_to_delete in subject.sections_images.all():
+                    object_to_delete.delete()
                 subject.number_of_sections -= 1
 
             subject.save()
